+++ MyMoney.Web/FrontEnd/React/src/state/session-slice.ts
