+++ MyMoney.Web/FrontEnd/React/src/state/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import sessionReducer from './sessionSlice';
+import sessionReducer from './session-slice';
 import { IAppState } from './types';
 
 export const store = configureStore<IAppState>({
